Fix all warnings

+++ File.cpp
@@ -25,7 +25,7 @@ bool FileArchive::Open(std::string filename, std::string directory)
     str.read((char*)&fat_size, 4);
 
 	str.seekg(fat_offset);
-	size_t pos = str.tellg();
+	size_t pos = (size_t)str.tellg();
     if(pos != (size_t)fat_offset)
         return false;
 
@@ -154,7 +154,7 @@ bool File::Open(std::string filename)
         myName = filename;
         myOffset = 0;
         myStream.seekg(0, std::ios::end);
-        myLength = myStream.tellg();
+        myLength = (size_t)myStream.tellg();
         myStream.seekg(0, std::ios::beg);
         myPosition = 0;
         myOpen = true;
@@ -200,7 +200,7 @@ uint32_t File::Read(void* cwhere, uint32_t size)
 
 	myStream.seekg(myOffset+myPosition, std::ios::beg);
     myStream.read((char*)cwhere, size);
-    uint32_t cnt = myStream.gcount();
+    uint32_t cnt = (uint32_t)myStream.gcount();
     myPosition += cnt;
 
     return cnt;
